@@ -3,19 +3,28 @@
   "version": "1.0.0",
   "description": "update new ticket trading information to LINE using LINE Notify API & LINE Messaging API",
   "engines": {
-    "node": "12.x"
+    "node": "18.x"
   },
-  "main": "index.js",
+  "main": "index.mjs",
+  "type": "module",
   "scripts": {
-    "start": "node index.js"
+    "start": "node index.mjs",
+    "dev": "nodemon index.mjs"
   },
   "dependencies": {
     "@line/bot-sdk": "^7.1.0",
-    "@waynechang65/ptt-crawler": "^2.1.0",
+    "dotenv": "^16.3.1",
     "express": "^4.17.1",
-    "node-fetch": "^2.6.1"
+    "node-fetch": "^2.6.1",
+    "nodemon": "^3.0.1",
+    "puppeteer": "^19.11.1"
   },
   "devDependencies": {
+    "eslint": "^8.50.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-n": "^16.1.0",
+    "eslint-plugin-promise": "^6.1.1",
     "prettier": "2.2.1"
   },
   "repository": {
