@@ -1,22 +1,21 @@
-const line = require("@line/bot-sdk");
-const express = require("express");
-const pttResult = require("./lib/crawler");
+import line from "@line/bot-sdk";
+import dotenv from "dotenv";
+import express from "express";
+import pttResult from "./lib/filter.mjs";
+dotenv.config();
 
-// create LINE SDK config from env variables
 const config = {
   channelAccessToken: process.env.LINE_CHANNEL_ACCESS_TOKEN,
   channelSecret: process.env.LINE_CHANNEL_SECRET,
 };
 
-// create LINE SDK client
 const client = new line.Client(config);
-
-// create Express app
-// about Express itself: https://expressjs.com/
 const app = express();
 
-// register a webhook handler with middleware
-// about the middleware, please refer to doc
+app.get("/health", (req, res) => {
+  res.send("ok");
+});
+
 app.post("/callback", line.middleware(config), (req, res) => {
   Promise.all(req.body.events.map(handleEvent))
     .then((result) => res.json(result))
